feat: add Alembic migration setup and integrate user_id into thoughts model

+++ backend/src/thoughts/controller.py
@@ -2,13 +2,15 @@ from src.thoughts.schema import thought_schema
 from sqlalchemy.orm import Session
 from src.thoughts.model import thought_model
 from fastapi import HTTPException
+from src.user.model import UserModel
 
-def create_thought(body: thought_schema,db:Session):
+def create_thought(body: thought_schema,db:Session,user: UserModel):
 
     data = body.model_dump()
 
     new_thought = thought_model(title=data["title"],
-                                content=data["content"])
+                                content=data["content"],
+                                user_id=user.id)
 
     db.add(new_thought)
     db.commit()
@@ -19,18 +21,21 @@ def create_thought(body: thought_schema,db:Session):
 
 
 
-def get_all_thought(db:Session):
-    thoughts = db.query(thought_model).all()
+def get_all_thought(db:Session,user: UserModel):
+    thoughts = db.query(thought_model).filter(thought_model.user_id == user.id).all()
     return thoughts
 
 
 
-def update_thought(body: thought_schema,thought_id:int,db:Session):
+def update_thought(body: thought_schema,thought_id:int,db:Session,user: UserModel):
 
-    thought = db.query(thought_model).get(thought_id)
+    thought: thought_model = db.query(thought_model).get(thought_id)
 
     if not thought:
         raise HTTPException(status_code=404,detail="Thought not found")
+
+    if thought.user_id != user.id:
+        raise HTTPException(status_code=401,detail="Unauthorized to update thought")
     
 
     body = body.model_dump()
@@ -46,11 +51,14 @@ def update_thought(body: thought_schema,thought_id:int,db:Session):
 
 
 
-def delete_thought(thought_id:int,db:Session):
-    thought = db.query(thought_model).get(thought_id)
+def delete_thought(thought_id:int,db:Session,user: UserModel):
+    thought: thought_model = db.query(thought_model).get(thought_id)
 
     if not thought:
         raise HTTPException(status_code=404,detail="Thought not found")
+
+    if thought.user_id != user.id:
+        raise HTTPException(status_code=401,detail="Unauthorized to delete thought")
 
     db.delete(thought)
     db.commit()
